Connected with all pages
Still quite a few bugs to be fixed.
From email (sent by @shreyabm) :
a) On the Question of the week page, the circular button should change colour into blue darken-2 when clicked on
b) The 'Are you sure you want to delete this question?' popup's layout is messed up and I don't have time to fix it.
c) In the 'Enter Question Details' popup, when subjective is clicked, there should be only 1 text area. If MCQ is selected, there should be 1+4 text areas. The code is all there. The onclick functionality needs to be adjusted to get the desired output.
d) Wherever you see teal, change it to blue darken-2 (I have done this as much as possible but I may have missed some)  and adjust the font sizes based on your discretion

Co-Authored-By: shreyabm <shreyabm@users.noreply.github.com>